move types to assets/src/types and update checkbox wording
